Automapper was added and seted up, code was little refactored, repo was changed

+++ src/Teams.Business/Services/ManageTeamsMembersService.cs
@@ -1,11 +1,9 @@
-using Microsoft.EntityFrameworkCore;
-/*using Microsoft.AspNetCore.Mvc;*/
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Teams.Data;
-using Teams.Data.Annotations;
-using Teams.Data.Models;
+using Teams.Business.Annotations;
+using Teams.Business.Models;
+using Teams.Business.Repository;
 using Teams.Security;
 
 namespace Teams.Business.Services
@@ -13,10 +11,11 @@ namespace Teams.Business.Services
     public class ManageTeamsMembersService : IManageTeamsMembersService
     {
         private readonly ICurrentUser _currentUser;
-        private readonly IRepository<Team, int> _teamRepository;
-        private readonly IRepository<TeamMember, int> _teamMemberRepository;
+        private readonly IRepository<TeamBusiness, int> _teamRepository;
+        private readonly IRepository<TeamMemberBusiness, int> _teamMemberRepository;
 
-        public ManageTeamsMembersService(IRepository<Team, int> teamRepository, IRepository<TeamMember, int> memberRepository, ICurrentUser currentUser)
+        public ManageTeamsMembersService(IRepository<TeamBusiness, int> teamRepository, 
+            IRepository<TeamMemberBusiness, int> memberRepository, ICurrentUser currentUser)
         {
             _currentUser = currentUser;
             _teamRepository = teamRepository;
@@ -25,43 +24,37 @@ namespace Teams.Business.Services
 
         public async Task<bool> RemoveAsync(int teamId, string memberId)
         {
-            var member = await _teamMemberRepository.GetAll()
-            .Where(x => x.MemberId == memberId && x.TeamId == teamId
-            && x.Team.TeamOwner == _currentUser.Current.Id()
-            && x.Team.TeamOwner != memberId)
-            .FirstOrDefaultAsync();
-            if (member != null)
-            {
-                return await _teamMemberRepository.DeleteAsync(member);
-            }
+            var allMembes = await _teamMemberRepository.GetAllAsync();
+            var member = allMembes.Where(x => x.MemberId == memberId && x.TeamId == teamId 
+                && x.Team.TeamOwner == _currentUser.Current.Id() && x.Team.TeamOwner != memberId).FirstOrDefault();
+            if (member != null) return await _teamMemberRepository.DeleteAsync(member);
             return false;
         }
 
         public async Task<bool> AddAsync(int teamId, string memberId)
         {
-            var alreadyInTeam = await _teamRepository.GetAll().
-                AnyAsync(t => t.TeamOwner == _currentUser.Current.Id() && t.Id == teamId && t.TeamMembers.Any(t => t.MemberId == memberId));
+            var allTeams = await _teamRepository.GetAllAsync();
+            var alreadyInTeam = allTeams.Any(t => t.TeamOwner == _currentUser.Current.Id() && t.Id == teamId 
+                && t.TeamMembers.Any(t => t.MemberId == memberId));
             if (!alreadyInTeam && memberId != _currentUser.Current.Id())
-            {
-                return await _teamMemberRepository.InsertAsync(new TeamMember { TeamId = teamId, MemberId = memberId });
-            }
+                return await _teamMemberRepository.InsertAsync(new TeamMemberBusiness { TeamId = teamId, MemberId = memberId });
             return false;
         }
 
-        public async Task<TeamMember> GetMemberAsync(int teamId, string memberId)
+        public async Task<TeamMemberBusiness> GetMemberAsync(int teamId, string memberId)
         {
-             return await _teamMemberRepository.GetAll()
-                    .Where(x => x.MemberId == memberId && x.TeamId == teamId)
-                    .FirstOrDefaultAsync();
+            var allMembers = await _teamMemberRepository.GetAllAsync();
+            var member = allMembers.Where(x => x.MemberId == memberId && x.TeamId == teamId).FirstOrDefault();
+            return member;
         }
 
-        public async Task<List<TeamMember>> GetAllTeamMembersAsync(int teamId, DisplayOptions options)
+        public async Task<List<TeamMemberBusiness>> GetAllTeamMembersAsync(int teamId, DisplayOptions options)
         {
-            var members = _teamMemberRepository.GetAll().Include(x => x.Member).Include(x => x.Team).ThenInclude(x => x.Owner).Where(x => x.TeamId == teamId);
+            var allMembers = await _teamMemberRepository.GetAllAsync();
+            var teamMembers = allMembers.Where(x => x.TeamId == teamId);
 
-            if (options.SortDirection == SortDirection.Ascending) return await members.OrderBy(x => x.Member.UserName).ToListAsync();
-
-            else return await members.OrderByDescending(x => x.Member.UserName).ToListAsync();
+            if (options.SortDirection == SortDirection.Ascending) return teamMembers.OrderBy(x => x.Member.UserName).ToList();
+            else return teamMembers.OrderByDescending(x => x.Member.UserName).ToList();
         }
     }
 }
